backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
mediapipe
opencv-python
librosa
soundfile
numpy
pandas
PyPDF2
python-docx
scikit-learn
